Switch from Autoconf to plain Make
GNU Autoconf is too complicated! Spek is a small program, so classic
Make will suffice. Now everything is in a short Makefile (<= 300 LOC),
which should be POSIX too. I've removed all the Autoconf leftovers I
could find, but maybe the po dir could be further cleaned.

The only thing left to do is to add a target to update the PO template
file with xgettext and merge it with existing translation files, using
msgmerge.